fix(android): prevent “flashing” activity on app start
This was due to a change made a while back to fix an issue with resetRootView.
The consequence was that we were waiting for the frame to be “attachedToWindow” to load the first frame page.
That PR make it so that we apply that check/fix only when the activity was reset. Thus it wont happen on app start